Refactor game detail data fetching and UI state
