+++ package.json
@@ -82,7 +82,6 @@
     "@types/node": "^22.19.17",
     "@uidotdev/usehooks": "^2.4.1",
     "@vidstack/react": "^1.12.13",
-    "@vitejs/plugin-react": "^4.7.0",
     "axios": "^1.15.1",
     "better-auth": "1.6.6",
     "better-sqlite3": "^12.9.0",
@@ -127,15 +126,15 @@
     "tailwindcss-animate": "^1.0.7",
     "tw-animate-css": "^1.4.0",
     "vaul": "^1.1.2",
-    "vite-plugin-pwa": "^1.2.0",
-    "workbox-window": "^7.4.0",
+    "vite-plugin-pwa": "^1.3.0",
+    "workbox-window": "^7.4.1",
     "zod": "^4.4.0"
   },
   "devDependencies": {
     "@playwright/test": "^1.59.1",
-    "@tailwindcss/postcss": "^4.2.2",
-    "@tailwindcss/vite": "4.0.7",
-    "@tanstack/devtools-vite": "^0.5.5",
+    "@tailwindcss/postcss": "^4.3.0",
+    "@tailwindcss/vite": "^4.3.0",
+    "@tanstack/devtools-vite": "^0.7.0",
     "@tanstack/nitro-v2-vite-plugin": "^1.154.9",
     "@tanstack/react-devtools": "^0.9.13",
     "@tanstack/react-router-ssr-query": "^1.166.11",
@@ -146,8 +145,9 @@
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
     "@typescript/native-preview": "7.0.0-dev.20260414.1",
-    "@vitejs/plugin-react-oxc": "^0.2.3",
+    "@vitejs/plugin-react": "^6.0.2",
     "autoprefixer": "^10.5.0",
+    "esbuild": "^0.27.7",
     "glob": "^11.1.0",
     "lint-staged": "^15.2.10",
     "nitro": "npm:nitro-nightly",
@@ -155,14 +155,13 @@
     "oxlint": "^1.60.0",
     "postcss": "^8.5.9",
     "simple-git-hooks": "^2.11.1",
-    "tailwindcss": "^4.2.2",
+    "tailwindcss": "^4.3.0",
     "typescript": "^5.9.2",
     "unenv": "^1.10.0",
-    "vite": "npm:rolldown-vite@7.1.9",
-    "vite-tsconfig-paths": "^5.1.4",
-    "workbox-precaching": "^7.4.0",
-    "workbox-routing": "^7.4.0",
-    "workbox-strategies": "^7.4.0",
+    "vite": "^8.0.13",
+    "workbox-precaching": "^7.4.1",
+    "workbox-routing": "^7.4.1",
+    "workbox-strategies": "^7.4.1",
     "wrangler": "3.99.0"
   },
   "packageManager": "pnpm@10.30.3",
