+++ pnpm-workspace.yaml
@@ -15,7 +15,6 @@ onlyBuiltDependencies:
   - workerd
 overrides:
   solid-js: 1.9.9
-  vite: npm:rolldown-vite@7.1.9
   '@better-auth/sso@<=1.6.0': 1.6.0
   workbox-build@<=7.4.1: 7.4.1
   nitropack@<=2.13.2: 2.13.2
